Add prospective employee model test
