+++ test/models/prospective_student_test.rb
@@ -3,10 +3,7 @@
 require 'test_helper'
 
 class ProspectiveStudentTest < ActiveSupport::TestCase
-  extend Support::Minitest::AssociationHelper
-  extend Support::Minitest::CallbackHelper
-  extend Support::Minitest::EnumerationHelper
-  extend Support::Minitest::ValidationHelper
+  include ProspectiveTest
 
   # relations
   belongs_to :academic_term
@@ -14,28 +11,16 @@ class ProspectiveStudentTest < ActiveSupport::TestCase
   belongs_to :language, optional: true
   belongs_to :student_entrance_type
   belongs_to :student_disability_type, optional: true
-  belongs_to :unit
 
   # validations: presence
   validates_presence_of :expiry_date
-  validates_presence_of :first_name
-  validates_presence_of :gender
-  validates_presence_of :id_number
-  validates_presence_of :last_name
-
-  # validations: uniqueness
-  validates_uniqueness_of :id_number
 
   # validations: length
   validates_length_of :address
-  validates_length_of :email
   validates_length_of :fathers_name
-  validates_length_of :first_name
   validates_length_of :high_school_code
   validates_length_of :high_school_branch
   validates_length_of :home_phone
-  validates_length_of :last_name
-  validates_length_of :mobile_phone
   validates_length_of :mothers_name
   validates_length_of :obs_registered_program
   validates_length_of :place_of_birth
@@ -45,13 +30,9 @@ class ProspectiveStudentTest < ActiveSupport::TestCase
 
   # enums
   enum additional_score: { handicapped: 1 }
-  enum gender: { male: 1, female: 2 }
   enum nationality: { turkish: 1, kktc: 2, foreign: 3 }
   enum placement_type: { general_score: 1, additional_score: 2 }
   enum system_register_type: { manual: 0, bulk: 1 }
-
-  # callbacks
-  before_create :standardization
 
   # search
   test 'prospective_student is a searchable model' do
